requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=hd77b12b_0
aiofiles=22.1.0=py38haa95532_0
aiosqlite=0.18.0=py38haa95532_0
altair=5.2.0=pyhd8ed1ab_0
anyio=3.5.0=py38haa95532_0
argon2-cffi=21.3.0=pyhd3eb1b0_0
argon2-cffi-bindings=21.2.0=py38h2bbff1b_0
arrow-cpp=10.0.1=ha81ea56_1
asn1crypto=1.5.1=py38haa95532_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=23.1.0=py38haa95532_0
aws-c-common=0.6.8=h2bbff1b_1
aws-c-event-stream=0.1.6=hd77b12b_6
aws-checksums=0.1.11=h2bbff1b_2
aws-sdk-cpp=1.8.185=hd77b12b_1
babel=2.11.0=py38haa95532_0
backcall=0.2.0=pyhd3eb1b0_0
backports=1.0=pyhd8ed1ab_3
backports.zoneinfo=0.2.1=py38h91455d4_8
beautifulsoup4=4.12.2=py38haa95532_0
blas=1.0=mkl
bleach=4.1.0=pyhd3eb1b0_0
blinker=1.7.0=pyhd8ed1ab_0
boost-cpp=1.82.0=h59b6b97_2
bottleneck=1.3.5=py38h080aedc_0
brotli-python=1.0.9=py38hd77b12b_7
bzip2=1.0.8=he774522_0
c-ares=1.19.1=h2bbff1b_0
ca-certificates=2024.3.11=haa95532_0
cachetools=5.3.2=pyhd8ed1ab_0
certifi=2024.2.2=py38haa95532_0
cffi=1.16.0=py38h2bbff1b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=win_pyh7428d3b_0
cloudpickle=2.0.0=pyhd3eb1b0_0
colorama=0.4.6=py38haa95532_0
comm=0.1.2=py38haa95532_0
contourpy=1.1.1=pypi_0
cryptography=41.0.3=py38h89fc84f_0
cycler=0.12.1=pypi_0
debugpy=1.6.7=py38hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
defusedxml=0.7.1=pyhd3eb1b0_0
entrypoints=0.4=py38haa95532_0
et_xmlfile=1.1.0=py38haa95532_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.13.1=py38haa95532_0
fonttools=4.47.2=pypi_0
freetype=2.10.4=h546665d_1
gflags=2.2.2=ha925a31_0
gitdb=4.0.11=pyhd8ed1ab_0
gitpython=3.1.40=pyhd8ed1ab_0
glog=0.5.0=hd77b12b_0
grpc-cpp=1.48.2=hfe90ff0_1
icu=73.1=h6c2663c_0
idna=3.4=py38haa95532_0
importlib-metadata=6.0.0=py38haa95532_0
importlib_metadata=6.0.0=hd3eb1b0_0
importlib_resources=6.1.0=py38haa95532_0
intel-openmp=2023.1.0=h59b6b97_46320
ipykernel=6.25.0=py38h9909e9c_0
ipython=8.12.2=py38haa95532_0
ipython_genutils=0.2.0=pyhd3eb1b0_1
ipywidgets=8.0.4=py38haa95532_0
jbig=2.1=h8d14728_2003
jedi=0.18.1=py38haa95532_1
jinja2=3.1.2=py38haa95532_0
jpeg=9e=h2bbff1b_1
json5=0.9.6=pyhd3eb1b0_0
jsonschema=4.19.2=py38haa95532_0
jsonschema-specifications=2023.7.1=py38haa95532_0
jupyter=1.0.0=py38haa95532_8
jupyter_client=7.4.9=py38haa95532_0
jupyter_console=6.6.3=py38haa95532_0
jupyter_core=5.5.0=py38haa95532_0
jupyter_events=0.8.0=py38haa95532_0
jupyter_server=1.23.4=py38haa95532_0
jupyter_server_fileid=0.9.0=py38haa95532_0
jupyter_server_ydoc=0.8.0=py38haa95532_1
jupyter_ydoc=0.2.4=py38haa95532_0
jupyterlab=3.6.3=py38haa95532_0
jupyterlab_pygments=0.2.2=py38haa95532_0
jupyterlab_server=2.25.1=py38haa95532_0
jupyterlab_widgets=3.0.9=py38haa95532_0
kiwisolver=1.4.5=pypi_0
krb5=1.20.1=h5b6d351_1
lcms2=2.12=h2a16943_0
lerc=2.2.1=h0e60522_0
libboost=1.82.0=h3399ecb_2
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libclang=14.0.6=default_hb5a9fac_1
libclang13=14.0.6=default_h8e68704_1
libcurl=8.1.1=h86230a5_0
libdeflate=1.7=h8ffe710_5
libevent=2.1.12=h56d1f94_1
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libpq=12.15=h906ac69_1
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=he2ea4bf_2
libthrift=0.15.0=h4364b78_2
libtiff=4.3.0=h0c97f57_1
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=3.0.0=pyhd8ed1ab_0
markupsafe=2.1.1=py38h2bbff1b_0
matplotlib=3.7.4=pypi_0
matplotlib-inline=0.1.6=py38haa95532_0
mdurl=0.1.0=pyhd8ed1ab_0
mistune=2.0.4=py38haa95532_0
mkl=2023.1.0=h6b88ed4_46358
mkl-service=2.4.0=py38h2bbff1b_1
mkl_fft=1.3.8=py38h2bbff1b_0
mkl_random=1.2.4=py38h59b6b97_0
nbclassic=1.0.0=py38haa95532_0
nbclient=0.8.0=py38haa95532_0
nbconvert=7.10.0=py38haa95532_0
nbformat=5.9.2=py38haa95532_0
nest-asyncio=1.5.6=py38haa95532_0
notebook=6.5.4=py38haa95532_1
notebook-shim=0.2.3=py38haa95532_0
numexpr=2.8.4=py38h7b80656_1
numpy=1.24.3=py38h79a8e48_1
numpy-base=1.24.3=py38h8a87ada_1
olefile=0.47=pyhd8ed1ab_0
openjpeg=2.4.0=hb211442_1
openpyxl=3.0.10=py38h2bbff1b_0
openssl=3.2.0=hcfcfb64_1
orc=1.7.4=h623e30f_1
oscrypto=1.2.1=pyhd3eb1b0_0
packaging=23.1=py38haa95532_0
pandas=2.0.3=py38h4ed8f06_0
pandocfilters=1.5.0=pyhd3eb1b0_0
parso=0.8.3=pyhd3eb1b0_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=8.3.2=py38h794f750_0
pip=23.3=py38haa95532_0
pkgutil-resolve-name=1.3.10=py38haa95532_0
platformdirs=3.8.1=py38haa95532_0
plotly=5.18.0=pypi_0
ply=3.11=py38_0
prometheus_client=0.14.1=py38haa95532_0
prompt-toolkit=3.0.36=py38haa95532_0
prompt_toolkit=3.0.36=hd3eb1b0_0
protobuf=3.20.3=py38haa244fe_1
psutil=5.9.0=py38h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=10.0.1=py38haa45b5f_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodomex=3.15.0=py38h2bbff1b_0
pydeck=0.8.0=pyhd8ed1ab_0
pygments=2.15.1=py38haa95532_1
pyjwt=2.4.0=py38haa95532_0
pyopenssl=23.2.0=py38haa95532_0
pyparsing=3.1.1=pypi_0
pyqt=5.15.10=py38hd77b12b_0
pyqt5-sip=12.13.0=py38h2bbff1b_0
pysocks=1.7.1=py38haa95532_0
python=3.8.18=h1aa4202_0
python-dateutil=2.8.3+snowflake1=py38haa95532_1
python-dotenv=1.0.0=pyhd8ed1ab_1
python-fastjsonschema=2.16.2=py38haa95532_0
python-json-logger=2.0.7=py38haa95532_0
python-tzdata=2023.3=pyhd3eb1b0_0
python_abi=3.8=2_cp38
pytz=2023.3.post1=py38haa95532_0
pywin32=305=py38h2bbff1b_0
pywinpty=2.0.10=py38h5da7b33_0
pyyaml=6.0.1=py38h2bbff1b_0
pyzmq=23.2.0=py38hd77b12b_0
qt-main=5.15.2=h19c9488_10
qtconsole=5.5.0=py38haa95532_0
qtpy=2.4.1=py38haa95532_0
re2=2022.04.01=hd77b12b_0
referencing=0.30.2=py38haa95532_0
requests=2.31.0=py38haa95532_0
rfc3339-validator=0.1.4=py38haa95532_0
rfc3986-validator=0.1.1=py38haa95532_0
rich=13.7.0=pyhd8ed1ab_0
rpds-py=0.10.6=py38h062c2fa_0
send2trash=1.8.2=py38haa95532_0
setuptools=68.0.0=py38haa95532_0
sip=6.7.12=py38hd77b12b_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pyhd8ed1ab_0
snappy=1.1.9=h6c2663c_0
sniffio=1.2.0=py38haa95532_1
snowflake-connector-python=3.2.0=py38h4ed8f06_1
snowflake-snowpark-python=1.10.0=py38haa95532_0
sortedcontainers=2.4.0=pyhd3eb1b0_0
soupsieve=2.5=py38haa95532_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
streamlit=1.29.0=pyhd8ed1ab_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.3=pyhd8ed1ab_0
terminado=0.17.1=py38haa95532_0
tinycss2=1.2.1=py38haa95532_0
tk=8.6.13=h5226925_1
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=py38haa95532_0
tomlkit=0.11.1=py38haa95532_0
toolz=0.12.0=pyhd8ed1ab_0
tornado=6.3.3=py38h2bbff1b_0
traitlets=5.7.1=py38haa95532_0
typing-extensions=4.7.1=py38haa95532_0
typing_extensions=4.7.1=py38haa95532_0
tzdata=2023c=h71feb2d_0
tzlocal=5.2=py38haa244fe_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=1.26.18=py38haa95532_0
utf8proc=2.6.1=h2bbff1b_0
validators=0.22.0=pyhd8ed1ab_0
vc=14.2=h21ff451_1
vc14_runtime=14.36.32532=hdcecf7f_17
vs2015_runtime=14.36.32532=h05e6639_17
watchdog=3.0.0=py38haa244fe_1
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=py38_1
websocket-client=0.58.0=py38haa95532_4
wheel=0.41.2=py38haa95532_0
widgetsnbextension=4.0.5=py38haa95532_0
win_inet_pton=1.1.0=py38haa95532_0
winpty=0.4.3=4
xlsxwriter=3.1.1=py38haa95532_0
xz=5.4.2=h8cc25b3_0
y-py=0.5.9=py38hb6bf4ef_0
yaml=0.2.5=he774522_0
ypy-websocket=0.8.2=py38haa95532_0
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=py38haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
